+++ R/autoboost.R
@@ -42,7 +42,7 @@
 #'
 #' @return A numeric matrix with attributes.
 #' @family Selectboost functions
-#' @author Frederic Bertrand, \email{frederic.bertrand@@utt.fr}
+#' @author Frederic Bertrand, \email{frederic.bertrand@@lecnam.net}
 #' @references \emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Frédéric Bertrand, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Bioinformatics, 2020. \doi{10.1093/bioinformatics/btaa855}
 #' @seealso \code{\link{boost}}, \code{\link{fastboost}}, \code{\link{plot.selectboost}}
 #' @examples
@@ -371,7 +371,7 @@ autoboost<-function(X,Y,ncores=4,group=group_func_1,func=lasso_msgps_AICc,corrfu
 #'
 #' @return A numeric matrix with attributes.
 #' @family Selectboost functions
-#' @author Frederic Bertrand, \email{frederic.bertrand@@utt.fr}
+#' @author Frederic Bertrand, \email{frederic.bertrand@@lecnam.net}
 #' @references \emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Frédéric Bertrand, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Bioinformatics, 2020. \doi{10.1093/bioinformatics/btaa855}
 #' @seealso \code{\link{boost}}, \code{\link{autoboost}}, \code{\link{plot.selectboost}}
 #' @examples
@@ -845,7 +845,7 @@ fastboost <-
 #'
 #' @return A matrix with the results.
 #' @family Selectboost analyse functions
-#' @author Frederic Bertrand, \email{frederic.bertrand@@utt.fr}
+#' @author Frederic Bertrand, \email{frederic.bertrand@@lecnam.net}
 #' @references \emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Frédéric Bertrand, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Bioinformatics, 2020. \doi{10.1093/bioinformatics/btaa855}
 #' @seealso \code{\link{fastboost}}, \code{\link{autoboost}}
 NULL
@@ -898,7 +898,7 @@ force.non.inc = function(object){
 #'
 #' @return An invisible list.
 #' @family Selectboost analyse functions
-#' @author Frederic Bertrand, \email{frederic.bertrand@@utt.fr}
+#' @author Frederic Bertrand, \email{frederic.bertrand@@lecnam.net}
 #' @references \emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Frédéric Bertrand, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Bioinformatics, 2020. \doi{10.1093/bioinformatics/btaa855}
 #' @seealso \code{\link{fastboost}}, \code{\link{autoboost}}
 #' @examples
@@ -1062,7 +1062,7 @@ plot.selectboost <-
 #'
 #' @return A list with the results.
 #' @family Selectboost analyse functions
-#' @author Frederic Bertrand, \email{frederic.bertrand@@utt.fr}
+#' @author Frederic Bertrand, \email{frederic.bertrand@@lecnam.net}
 #' @references \emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Frédéric Bertrand, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Bioinformatics, 2020. \doi{10.1093/bioinformatics/btaa855}
 #' @seealso \code{\link{fastboost}}, \code{\link{autoboost}}
 NULL
@@ -1232,7 +1232,7 @@ summary.selectboost <-
 #'
 #' @return An invisible list.
 #' @family Selectboost analyze functions
-#' @author Frederic Bertrand, \email{frederic.bertrand@@utt.fr}
+#' @author Frederic Bertrand, \email{frederic.bertrand@@lecnam.net}
 #' @references \emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Frédéric Bertrand, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Bioinformatics, 2020. \doi{10.1093/bioinformatics/btaa855}
 #' @seealso \code{\link{fastboost}}, \code{\link{autoboost}} and \code{\link{summary.selectboost}}
 NULL
@@ -1285,7 +1285,7 @@ plot.summary.selectboost <- function(x, ...)
 #'
 #' @return An invisible list.
 #' @family Selectboost analyze functions
-#' @author Frederic Bertrand, \email{frederic.bertrand@@utt.fr}
+#' @author Frederic Bertrand, \email{frederic.bertrand@@lecnam.net}
 #' @references \emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Frédéric Bertrand, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Bioinformatics, 2020. \doi{10.1093/bioinformatics/btaa855}
 #' @seealso \code{\link{fastboost}}, \code{\link{autoboost}} and \code{\link{summary.selectboost}}
 NULL
@@ -1444,7 +1444,7 @@ trajC0.selectboost <-
 #'
 #' @family Selectboost analyze functions
 #'
-#' @author Frederic Bertrand, \email{frederic.bertrand@@utt.fr}
+#' @author Frederic Bertrand, \email{frederic.bertrand@@lecnam.net}
 #'
 #' @references \emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Frédéric Bertrand, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Bioinformatics, 2020. \doi{10.1093/bioinformatics/btaa855}
 #' @seealso \code{\link{fastboost}} and \code{\link{autoboost}}