+++ R/boost.R
@@ -6,7 +6,7 @@
 #'
 #' @return Various types depending on the function.
 #' @family Selectboost functions
-#' @author Frederic Bertrand, \email{frederic.bertrand@@utt.fr}
+#' @author Frederic Bertrand, \email{frederic.bertrand@@lecnam.net}
 #' @references \emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Frédéric Bertrand, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Bioinformatics, 2020. \doi{10.1093/bioinformatics/btaa855}
 #' @seealso \code{\link{fastboost}}, \code{\link{autoboost}}
 #' @examples