+++ R/group_func_2.R
@@ -14,7 +14,7 @@
 #'   \item "type": "normal"
 #'   \item "length.groups" the length of each groups.
 #' }
-#' @author Frederic Bertrand, \email{frederic.bertrand@@utt.fr}
+#' @author Frederic Bertrand, \email{frederic.bertrand@@lecnam.net}
 #' @references \emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Frédéric Bertrand, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Bioinformatics, 2020. \doi{10.1093/bioinformatics/btaa855}
 #' @seealso \code{\link{group_func_2}} \code{\link{boost.findgroups}}, \code{\link[igraph]{infomap.community}} and \code{\link[igraph]{igraph}}.
 #' @examples